api/requirements.txt
google-cloud-vision==3.4.5

